+++ demo/src/main/java/com/example/demo/repository/ComprasRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface ComprasRepository extends JpaRepository<ComprasEntity,Integer> {
+public interface ComprasRepository extends JpaRepository<ComprasEntity,Long> {
 }
